Expanded check for infrared beacon
The check for the beacon now takes to parts:

 - Have the two zones in question surpassed the average noise value?

 - Is the second zone (the one which will get signal second) nearly
   equal to the first zone or greater?

If both of these checks pass, the infrared beacon is considered found.
The first check is a persistant flag, so the point at which we are
"in range" is remebered by each iteration of the seeker loop.

Currently, this is implemented as a global; further patches should
hopefully change this.