Added ring divider
Added code to manage the ring divider servo. This required adding
lines to both the joystick 2 watchdog task and the manipulator
updator task.

Updated power levels for the first joystick's DPad input.

In an attempt to fight lag, we tried several things. Unfortunately,
none of them worked. The issue is probably more related to the speed
of RobotC than to our code. These changes are:

 - Staggered the task initialization.

 - Lowered the delay between joystick updates.